shortcuts_sync/requirements.txt
MySQL-python==1.2.5
mysql_connector_repackaged==0.3.1
pandas==1.5.0
pymongo==4.2.0
pyodbc==4.0.34
SQLAlchemy==1.4.41
typing_extensions==4.4.0
